Modify and add new combination button. (#70) (#100)
Neogo combination button "BCD" has been changed to "BC".

The "BC" key combination is difficult to press on an Xbox controller.
Especially when playing The King of Fighters 2002.

"BCD" combination button can be implemented by L1+L2, or L1+B, or L2+Y.

Add PGM combination button: R1-AB, R2-ABC, L1-CD, L2-BC.

Add CPS1 combination button: R2-AB.

The ABXY key of the controller is not suitable for pressing key combinations such as AB, CD, BC, ABC, BCD with the right thumb.
With the assistance of the shoulder buttons, the gaming experience of the controller will be better.
It's also a lot easier to play on Android with touchscreen buttons.